Add Name Displays (#29)
* Refactor reducers. Add getIsPlayersTurn

* Rename playerById to byId in playerReducer

* Make selectProps local function to selectors

* Create selectPlayerNameById cached selector

* Create NameDisplay component

* Add Glow to CurrentPlayer
