+++ front-end/src/selectors/gameStateSelectors.js
@@ -22,8 +22,9 @@ import {
   TRANSFER_MOVE,
 } from '../constants/gameMoves';
 import { selectDiscardPileLength } from './cardSelectors';
-import { selectYourId } from './index';
+import { selectYourId } from './playerSelectors';
 
+const selectProps = (_, props) => props;
 export const selectCurrentGameState = (state) => state.gameState.currentState;
 const selectPutAllowed = (state) => state.gameState.putAllowed;
 export const getCanPut = createSelector(
