+++ front-end/src/selectors/index.js
@@ -1,19 +1,11 @@
 import { createSelector } from 'reselect';
 import createCachedSelector from 're-reselect';
-const selectGameState = (state) => state.gameState;
-
 export * from './cardSelectors';
+export * from './playerSelectors';
+export * from './gameStateSelectors';
+export * from './turnSelectors';
 
 export const selectRoundState = (state) => state.round.currentState;
-
-export const selectYourId = (state) => state.players.yourId;
-export const selectAllPlayers = (state) => state.players.allPlayers;
-export const selectOpponentId = createSelector(
-  selectYourId,
-  selectAllPlayers,
-  (yourId, allPlayers) => allPlayers.filter((ids) => ids !== yourId).pop()
-);
-const selectPlayer = (id) => (state) => state.players.byId[id];
 const selectPlayerReady = (state) => state.round.ready.playerReady;
 const selectTotalReady = (state) => state.round.ready.totalReady;
 export const selectReady = createSelector(
