+++ front-end/src/selectors/turnSelectors.js
@@ -1,6 +1,7 @@
-import { selectYourId } from './index';
+import { selectYourId } from './playerSelectors';
 import { createSelector } from 'reselect';
 
+const selectProps = (_, props) => props;
 export const selectCurrentPlayer = (state) => state.turn.currentPlayerTurn;
 export const selectCurrentPuttingPlayer = (state) => state.turn.currentPuttingPlayer;
 export const getIsYourTurn = createSelector(
@@ -12,4 +13,10 @@ export const getAreYouCurrentPuttingPlayer = createSelector(
   selectYourId,
   selectCurrentPuttingPlayer,
   (yourId, puttingPlayer) => yourId === puttingPlayer
+);
+
+export const getIsPlayersTurn = createSelector(
+  selectCurrentPlayer,
+  selectProps,
+  (currentPlayer, props) => currentPlayer === props.playerId
 );
